+++ packages/vseed/src/pipeline/advanced/chart/pipes/config/config.ts
@@ -91,3 +91,5 @@ export const donutConfig: AdvancedPipe = pieConfig
 export const roseConfig: AdvancedPipe = pieConfig
 export const roseParallelConfig: AdvancedPipe = pieConfig
 export const funnelConfig: AdvancedPipe = pieConfig
+export const heatmapConfig: AdvancedPipe = pieConfig
+export const radarConfig: AdvancedPipe = pieConfig