+++ packages/vseed/src/pipeline/spec/chart/pipes/axes/angle.ts
@@ -7,7 +7,7 @@ export const angleAxis: SpecPipe = (spec, context) => {
   const { dimensions, measures } = advancedVSeed
 
   if (!dimensions || !measures) {
-    return spec
+    return result
   }
 
   const is0D = dimensions.length === 0