+++ packages/vseed/src/types/properties/chartType/chartType.ts
@@ -3,6 +3,7 @@ import { z } from 'zod'
 export type ChartType =
   | 'table'
   | 'pivotTable'
+  // cartesian
   | 'line'
   | 'column'
   | 'columnPercent'
@@ -13,17 +14,22 @@ export type ChartType =
   | 'area'
   | 'areaPercent'
   | 'areaRange'
+  | 'scatter'
+  | 'dualAxis'
+  // polar
   | 'rose'
   | 'roseParallel'
   | 'pie'
   | 'donut'
-  | 'dualAxis'
-  | 'scatter'
+  | 'radar'
+  // other
+  | 'heatmap'
   | 'funnel'
 
 export enum ChartTypeEnum {
   Table = 'table',
   PivotTable = 'pivotTable',
+  // cartesian
   Line = 'line',
   Column = 'column',
   ColumnPercent = 'columnPercent',
@@ -34,18 +40,23 @@ export enum ChartTypeEnum {
   Area = 'area',
   AreaPercent = 'areaPercent',
   AreaRange = 'areaRange',
+  DualAxis = 'dualAxis',
+  Scatter = 'scatter',
+  // polar
   Rose = 'rose',
   RoseParallel = 'roseParallel',
   Pie = 'pie',
   Donut = 'donut',
-  DualAxis = 'dualAxis',
-  Scatter = 'scatter',
+  Radar = 'radar',
+  // other
   Funnel = 'funnel',
+  Heatmap = 'heatmap',
 }
 
 export const zChartType = z.enum([
   'table',
   'pivotTable',
+  // cartesian
   'line',
   'column',
   'columnPercent',
@@ -56,11 +67,15 @@ export const zChartType = z.enum([
   'area',
   'areaPercent',
   'areaRange',
+  'scatter',
+  'dualAxis',
+  // polar
   'rose',
   'roseParallel',
   'pie',
   'donut',
-  'dualAxis',
-  'scatter',
+  'radar',
+  // other
   'funnel',
+  'heatmap',
 ])